consolidate tagging status into folderSlice and fix progress bar layout
